Change structure for next PRs
