+++ index/src/lib.rs
@@ -1,14 +1,11 @@
 #![deny(bare_trait_objects)]
 
-extern crate exocore_common;
-extern crate exocore_data;
-extern crate exocore_transport;
-
 #[macro_use]
 extern crate failure;
 
-extern crate tantivy;
-
 pub mod entity;
+pub mod errors;
+pub mod index;
+pub mod mutations;
 pub mod queries;
 pub mod store;
